+++ src/Components/GuestManagement/Local.jsx
@@ -9,6 +9,7 @@ import { Link } from "react-router-dom";
 import moment from "moment";
 import FilterModal from "../Common/Filter/FilterModal";
 import StatusFilter from "../Common/Filter/StatusFilter";
+import Swal from "sweetalert2";
 
 const Local = () => {
   const [filterFields, setFilterFields] = useState({
@@ -164,9 +165,20 @@ const Local = () => {
 
   const handleDeleteItem = async (id) => {
     try {
-      let { data } = await DeleteGuide(id);
-      if (data && !data.error) {
-        getGuides();
+      let confirm = await Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      });
+      if (confirm?.isConfirmed) {
+        let { data } = await DeleteGuide(id);
+        if (data && !data.error) {
+          getGuides();
+        }
       }
     } catch (error) {
       console.log(error);
